requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-image-compressor==0.3.6
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
filelock==3.9.0
identify==2.5.18
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
nodeenv==1.7.0
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.1.0
pre-commit==3.1.1
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.20.0
